atualizacao funçoes de geracao de populacao, fitness, selecao, crossover

+++ src/main/java/com/isonoe/cardeal/model/WayPointLink.java
@@ -7,6 +7,6 @@ import java.math.BigDecimal;
 
 @Data
 public class WayPointLink {
-    BigDecimal id;
+    Integer id;
     BigDecimal distance;
 }
